requirements/base.txt
django-reversion==1.8.0
Django==1.6
South==0.8.2
